Fix building on Linux with RMT_USE_OPENGL
Because GL/glx.h isn't #included, it doesn't know about
glXGetProcAddressARB(). Including it causes conflicts with other
definitions like the GLint typedefs, so I just added a declaration.

For the same reason (system GL headers not included). GLAPI isn't known,
so it tried to use WINGDIAPI which is windows-specific, so fall back to
extern on non-_WIN32.