Avoid asserting when the WebSocket server could not be created
If for instance the port is already occupied, `Remotery_Constructor`
will return with an error before g_Remotery and g_RemoteryCreated could
be set.

This situation triggers an assertion immediately in `Remotery_Destructor`
since it expects these globals to have been set.

This fixes it by asserting only when coming from the intended API entry
point, `rmt_DestroyGlobalInstance` which is where the assertion would be
most useful.