Bugfix: When resizing HashTable, the key would be copied instead of the value
This caused crashes later on because StringTable_Find would return the
number of the key instead of the value (which is an offset to a
pointer), resulting in out of bound reads.

This bug would trigger if a sample changed its name dynamically between
frames at a very fast rate.